+++ instruments.go
@@ -19,6 +19,7 @@ import (
 	"strings"
 
 	"go.opentelemetry.io/otel/metric"
+	semconv "go.opentelemetry.io/otel/semconv/v1.30.0"
 )
 
 const (
@@ -36,21 +37,32 @@ type dbStatsInstruments struct {
 }
 
 type instruments struct {
-	// The latency of calls in milliseconds
-	latency metric.Float64Histogram
+	// The legacyLatency of calls in milliseconds
+	legacyLatency metric.Float64Histogram
+	// The duration of calls in seconds
+	duration metric.Float64Histogram
 }
 
 func newInstruments(meter metric.Meter) (*instruments, error) {
 	var instruments instruments
 	var err error
 
-	if instruments.latency, err = meter.Float64Histogram(
+	if instruments.legacyLatency, err = meter.Float64Histogram(
 		strings.Join([]string{namespace, "latency"}, "."),
 		metric.WithDescription("The latency of calls in milliseconds"),
 		metric.WithUnit("ms"),
 	); err != nil {
-		return nil, fmt.Errorf("failed to create latency instrument, %v", err)
+		return nil, fmt.Errorf("failed to create legacy latency instrument, %v", err)
 	}
+
+	if instruments.duration, err = meter.Float64Histogram(
+		semconv.DBClientOperationDurationName,
+		metric.WithDescription(semconv.DBClientOperationDurationDescription),
+		metric.WithUnit(semconv.DBClientOperationDurationUnit),
+	); err != nil {
+		return nil, fmt.Errorf("failed to create duration instrument, %v", err)
+	}
+
 	return &instruments, nil
 }