+++ rows_test.go
@@ -88,7 +88,7 @@ func TestOtRows_Close(t *testing.T) {
 					ctx, sr, tracer, _ := prepareTraces(false)
 
 					mr := newMockRows(tc.error)
-					cfg := newMockConfig(t, tracer)
+					cfg := newMockConfig(t, tracer, nil)
 					cfg.SpanOptions.SpanFilter = spanFilterFn
 
 					// New rows
@@ -159,7 +159,7 @@ func TestOtRows_Next(t *testing.T) {
 					ctx, sr, tracer, _ := prepareTraces(false)
 
 					mr := newMockRows(tc.error)
-					cfg := newMockConfig(t, tracer)
+					cfg := newMockConfig(t, tracer, nil)
 					cfg.SpanOptions.RowsNext = tc.rowsNextOption
 					cfg.SpanOptions.SpanFilter = spanFilterFn
 
@@ -243,7 +243,7 @@ func TestNewRows(t *testing.T) {
 							ctx, sr, tracer, dummySpan := prepareTraces(tc.noParentSpan)
 
 							mr := newMockRows(false)
-							cfg := newMockConfig(t, tracer)
+							cfg := newMockConfig(t, tracer, nil)
 							cfg.SpanOptions.OmitRows = omitRows
 							cfg.SpanOptions.SpanFilter = spanFilterFn
 							cfg.AttributesGetter = tc.attributesGetter